hakyll/default.html: disable compilation of all double-quotes into <q></q> tags for syntax-highlighting/shading given mixed feedback; they can be added manually for select quotes
This is a regular bone of contention with Obormot and several other people, and feedback is highly split on whether the gwern.net quote-highlighting is a good thing: https://twitter.com/gwern/status/1100109285483266050 88 votes total. 34% no, 52% yes, 14% subtler, +1 positive comment on /r/gwern. It also is unusual, fragile, complex (CSS+JS+Pandoc), making it a perennial problem in site design changes, and also somewhat misleading as I often use double-quotes for things which are, strictly speaking, not quoting anyone or anything (rhetorical quotes, paraphrases, scare-quotes, etc). Given all this, it makes more sense to keep it off by default, and then write <q>'quote'</q> manually for quotes which ought to be highlighted for the reader as important or literal quotes but not large enough to be blockquotes.