initial.css: fix 'trapped scrolling' behavior on code blocks on mobile (particularly obnoxious on /GPT-2)
this was deliberate: so people don't accidentally scroll the code block off the screen as they're trying to scroll to the end of it. Which is an issue, but, obviously, not as big an issue as not being able to deliberately scroll it off the screen (which is obvious in retrospect...) Not a good idea, so remove it.