Typography: image inversion: add a manual review pass for inverted images as well per https://twitter.com/skullpile5/status/1334990881485541377
For image inversion, false positives are worse than false negatives; even invertible images, when grayscaled+dimmed for dark mode, look fine (or at worse, a little bit awful), but if an image gets inverted which shouldn't be, it tends to look a lot awful. So we want to check any new image which the heuristic fires for, in case we need to override it.