WE: bite the bullet and switch to the Mobile HTML API (turns out it has an section endpoint too, thank goodness); nicely simplifies and greatly increases correctness of extracts thus far...
On an amusing sidenote, this is now the *sixth* WP popups implementation for gwern.net: We had the version which queried the API for each link on page load; then the one which did on demand popups; then the Page Preview backend one  (which fed into the multiple implementations of popups from inlined to recursive to external-bibliography to external fragments), then the old API MediaWiki source, then the Mobile HTML (parsed into sections), and now the Mobile introduction-only HTML one. I hope this one sticks.