LA: +override feature for specific links
background: https://twitter.com/gwern/status/1380680645396021250 This is groundwork for switching all gwern.net WP links to the WP history instead, because so many WP articles keep decaying steadily under deletionist assault (and in at least some cases, this seems to be *because* they were linked by me). Should we do that, however, there will be WP links we'd want to always go to the live version, perhaps because it's too fast-moving, and the whitelist is an unpleasant way to do that.