gB/LM: +page-defined tags - some pages are link-dumps and de facto 'tags'
using the backlinks database, certain pages cause any link on them to get a certain tag. so now everything on /Sunk-costs will automatically get a 'sunk-cost' tag (ie. have sunk-cost tag on their popups and appear in the /docs/sunk-cost/index page). this avoids the tedium of going through, say, 500 links on /GPT-3 and manually annotating each with 'ai/gpt' and so on and so forth