gD: check for an annotation on the tag-directory to use as an introduction/preface/abstract
wouldn't it be nice to be able to put in a brief discussion or preface into link-directories, like some of the notes/* pages do? well, why not have generateDirectory.hs check for an annotation for the directory index and substitute it in if it exists. so to turn https://www.gwern.net/notes/Sparsity into a '/docs/ai/sparsity/index' tag page, I tag all the links with 'ai/sparsity' and turn the introduction into an annotation for the link '/docs/ai/sparsity/index'...