LM: pages to tags inference: check for hashless URLs
I noticed '/notes/Scaling' wasn't setting the 'ai/scaling' tag on a *lot* of links, which is a bug; turns out they are all links with '#' in them, mostly for affiliation, and the lookup call on the URL stripped '#' for consistency, except with no fallback (like we needed in generating directory indexes for similar reasons). So I add a fallback.