default.html: disable dropcaps on index pages, particularly directory-tags & link-bibliographies
This fixes the occasional complaint about 'the arrow is too big': they were on Chrome, and were looking at directory-tags. This doesn't happen on FF, but on Chrome (narrow definition of letters?), the initial '↑' for parent directory (inside a <p>, not a list) gets treated as ::first-letter and gets drop-capped, turning into a *giant* arrow (there being no dropcap defined for '↑' ofc), which looks terrible. So this patch disables dropcaps on index pages (which includes directory-tags & link-bibliographies) by switching on '$' and disabling the cssExtension logic (currently used only for dropcaps, and at this rate, always will be dropcap-only).