embed.hs: switch to Data.Binary, which scales to >6.4k embeddings while reading quickly and without erroring out
puzzled why my initial effort failed with a bunch of inscrutable type errors, depressing me and leading to a bunch of procrastination, and retrying tonight, it Just Worked. it isn't as much smaller as I hoped, but it does read fast and without bombing out in stack overflow errors. so now I can move on to the cosine-similarity search part.