dark-mode: for .invertible-not, do not even grayscale because the color is really important (or else it would not be manually marked up as .invertible-not)
Prompted by /Red: if we didn't .invertible-not the images there, most would be inverted (often being only 2 or 3 colors on a white background); however, defaulting to grayscale is also not ideal - the vivid colors are a critical part of /Red. What do? well, most images are handled by the heuristics or explicit .invertible classes. If we set a .invertible-not, that means we care a lot about colors. So we make .invertible a *third* state: neither inverted nor grayscaled. So now the hierarchy goes .invertible > .invertible-auto > default > .invertible-not.