LA: automatically filter out gwern.net or local links for convenience when updating broken links
Currently, everytime I fix a broken link by hosting it locally, this often means editing archive.hs to delete it, since it will be in there, but of course, LA is irrelevant to local links, and that triggers warnings. So let's just drop those automatically, assuming they were intentional & related to link-fixing.