annotations: date-handling: delete YYYY-01-01 dates entirely as bullshit; compress date display to just YYYY in popins/popups
I've found 'YYYY-01-01' dates to be so unreliable as to not be worth displaying in any context. TO a lesser extent, -MM-DD dates tend to be both unreliable and not that useful to readers; between vertical space consumption & unreliability, not worth prioritizing/highlighting, so we stash the (supposed) 'date' in a tooltip for anyone who really cares, and display it only on tag-directories and link-bibliographies where there's a ton of space and the reader can be presumed to be much more interested in the details like the full list of authors or the full date etc.