hakyll: generate titlePlain for all pages automatically instead of manually adding it to YAML metadata on pages that need it
the manual approach is just too error-prone - I occasionally forget it, like on /DNB-FAQ, and it leads to raw markup being exposed to users & screwed up titles which don't get detected by any existing lints. so let's just do away with it and create it always, not expensive enough to care about.