LM: begin restoring gwern.net abstract annotations
We are backtracking on live popups of gwern.net pages because a number of users on HN, LW, and elsewhere have complained that live page popups, particularly for pages like /Lorem or /GPT-3, are a bad user experience because they can take up to 15s to load, with no feedback, and are painfully unpredictable. So we backtrack to the previous implementation: we scrape the abstracts for top-level pages like '/Lorem', and provide that as an annotation. Links like '/Lorem#images', which we can't auto-write abstracts for (and there may be no sensible abstract or annotation one *could* write), will remain live link popups.