LM: gwern.net: define a new class 'abstractNot', to disable section scraping of abstracts
This is for sections which are placeholders for essays that got split out. They are styled as 'div class=abstract', but that leads to spurious annotations for URLs like https://www.gwern.net/Turing-complete#how-many-computers-are-in-your-computer which are irrelevant and redundant with their new home https://www.gwern.net/Computers ; so we set an additional class, 'abstractNot', on those abstract blocks, and the scraper checks for that class and if it's present, abandons the section scrape. So now /Turing-complete will work, but /Turing-complete#how-many will just fail because there's no useful abstract there.