*.hs: rename backlinks-not -> backlink-not for consistency
'backlinks-not' sorta implies that it disables backlinks for a URL (either the one in the Link or the URL that link is within), because it's plural and refers to 'backlink*s*'; ofc, it doesn't, it merely says to not include *this* singular instance. so it's more like 'don't this one backlink', so removing the 's' adds a bit of consistency & predictability.