LM: partials: try to parse a bit of metadata (title/authors/date) out of tooltips where possible, to get more useful partials
This has the additional benefit of avoiding partial popups *degrading* user experience. Right now, a lot of links have tooltips written with their title, or possibly title/author/date, because that is easy (or they long predate the annotation functionality). But with partial popups, the regular browser tooltip is suppressed. This means that something like 'The [sorrow](http://thecodelesscode.com/case/96 "The Codeless Code: Case 96: Stateless")' actually gets worse for the reader, because before they could mouse over, get the tooltip with the title, and that was all the metadata I had to provide them; with the partial popup, that is hidden and they are stuck with just a backlink & tag. But if we pull the title from the tooltip when we create the annotation, then that title will be in the partial and it's fine.