gD: deprecate 'tag-directory' terminology for just 'tag'; rm file prefix from local docs as no longer relevant
Simplify by hiding the tag/directory conflation. A reader doesn't care about whether a doc is remote or local, or what directory it is in, as long as they work equally well (and things like the link archives means there's no hard-and-fast distinction in the first place). We should only show the filepath in tag indexes if we have no better metadata to show.