LM: partials: further tweaks to writing out partials
Obormot will explore putting a '(i)' or '(...)' or similar icon on links which should pop up literally as live-links, such as images or PDFs or live-links, which provides the 'partial' information, as a compromise between displaying the 'partial' metadata in a wrapper around the live popup (which he thinks would take up unacceptably much room on most displays). So we need to actually write out thos partial HTML fragments for each such link. This change may not be quite right but let's see how it goes...