LM: annotations: backlinks appendage: strictly the transclusion of backlinks
Since the backlinks will always be transcluded despite being lazy, because annotations are so short, do strict transclusion (.include-instant). This may or may not yield any perceptible latency or UI improvement (since as soon as the annotation popups or transcludes, the laziness ought to fire immediately so the strictness may not matter), but it does document intent, and we don't use .include-instant anywhere right now so this is a good place to exercise/test the functionality.