LA/LM: annotations: push LinkAuto rewrite into annotation creation, not compilation, for compile spedup & clarity of what gets linked
The cost to run LA on each annotation every rebuild, despite LA changing rarely, is a major part of site compile time. But there's an additional downside: when I am reading or editing the original HTML/YAML, I don't know what, if anything, is going to be added by LA. For annotations which got created by hand and passed through preprocess-markdown.hs as of a few months ago, I see the LAs, but that doesn't apply to auto-created annotations such as Arxiv. So, I save both ludicrous amounts of compile-time *and* get a better idea what missing LAs are needed if I just bite the bullet and run LA on new annotations. Then when new LAs are added, I can simply rerun LA on the corpus to update it once & for all.