LM: partial annotations: expose partials correctly by checking status by brute force
partials are still badly broken because hasAnnotation doesn't infer partial status correctly (it depends on more metadata than exists inside the current MetadataItem data type, like the existence of backlinks/similar-links). so rather than spend a lot of time threading around the back/similarlinks databases & attempt to duplicate the partials logic of writeAnnotationFragment, for now I hack around the issue by hitting it with the big heavy rock of unsafePerformIO to check the existence of an annotation on-disk. If a link doesn't satisfy the requirements for a full annotation but there's a HTML snippet for it anyway, then it must be a partial, lol. In the long run, merging the different metadata databases is the solution.