annotations: WP styling: per anonymous feedback, nuke .sidebar class
the .sidebar class is responsible for all the 'a series in X' infobox-like lists of scores or hundreds of entries which require an absurd amount of scrolling and are extremely obnoxious both in popups & tag-directories. a check of WP articles in 40 tags showed ~15 instances of .sidebar which were the obnoxious kind, and 0 desirable-here. We could hypothetically try to match on 'a series in' except there turn out to be exceptions like https://www.gwern.net/docs/psychology/novelty/index#arousal where that would fail. So, not worth trying merely because of unknown hypothetical false positives.