LM: partial: fix subtle conditional bug in WP partials
any tagged WP link (making it a partial) was not being transcluded in link-bibliographies/tag-directories because it got both .link-annotated and .link-annotated-partial, while .link-annotated-partial kills .include-annotation (because -partial *means* no full annotation exists... except the WP 'annotations' are 'virtual' and pulled from WP). Example bug on /Problem-14 now: 'Optimal stopping' is a bare link in the link-bibliography (because I happened to tag it 'statistics/probability'), but 'Random walk' is a normal WP transclusion (I didn't bother to tag it). Just looks broken to a reader. Fix: so inside 'addHasAnnotation' we add another conditional to avoid ever setting a -partial on a WP link.