LM: annotation metadata: rm the href # self-link but keep the ID
self-links inside the annotation itself (rather than as default rewritten to pop up inside the 'logical page') turn out to be a very complicated idea. Instead of enabling that, we will leave the bl/sl links pointing to the fragments as before, put a unique ID on the bl/sl transclusions, and specialcase a scroll-to-transcluded-section for popins. UI-wise, this is not too important for popups, where popping up the fragments is an acceptable experience. (So here I am surfing along and I tap 'backlink' and I get an entire popin with a single link in it I now have to X my way out of to resume reading for popins, it seems like it would be lower friction to either transclude-right-there the bl/sl when I tap on it, at which point I either tap an interesting link to pop it in as normal, or I shrug and scroll with my finger a little more. A second possibility: tapping on the bl/sl at the top of the pop-in then scrolls down to the bl/sl transclude; and then if I regret that, I just shrug and scroll up with my finger back to the top or middle or wherever looks interesting.)