Typography: image lazy loading: enable basic support for overriding using loading='eager'
Forcing loading='lazy' on *all* images is now unnecessary/harmful. Specifically in annotations, because annotations are either popped up or lazy-transcluded, and in both cases we *want* the images strictly loaded (particularly if we are rendering a popup 'offscreen' to reduce latency, where lazy images defeat part of the point).