transcludes: strip the .include-spinner-not class, as it is increasingly verbose & used everywhere; Obormot will make the spinner more contextual.
Transclude loading spinnies should default to off for most transcludes. it seems I am constantly sprinkling around '.include-spinner-not' to disable it, and that almost all the time that I do see a spinnie, it is extremely brief & distracting, not actually useful in telling me 'be patient, it's not broken, just going to take a while' - thus far, it either happens within a split second or it's completely broken and doesn't happen at all...
Spinnies should be more contextual: transcluding an entire page? yeah, then default to spinnie 'on'. transcluding an anchor? mm... maybe default to 'off'? transclude an annotation? definitely default to 'off'.
The classes '.include-spinner' and '.include-spinner-not' will of course be available  as overrides for specific instances.