sync: syntax highlighting: strip self-links
Pandoc skylighting inserts self-links into source code to make them linkable; this interacts very badly with our CSS so we forked skylighting to delete that (it's not configurable). Which is fine, except... apparently my Pandoc binary now links a *different* library? so it now inserts them into the syntax-highlighted files (but not actual Hakyll-compiled pages). We will hack around it by sed, as usual. (If this works out, we may apply this to regular pages too to defeat the skylighting self-linking instead of relying on a local fork, which as this illustrates is fragile.)