pop-frame title: rm '[ARCHIVED]' entirely as redundant/hard to get right
putting it suffix raises problems of its own: what happens with *long* titles? since I don't like it at the beginning and it's problematic at the end, and strategies like truncating titles are increasingly complex and unsatisfactory, worth asking, why have it at all? the archive link will be in the annotation itself, after all. no good answer right now, so let's just delete it and see if we miss it. (We probably won't.)