LM: annotation formatting: title: delete curly-quotes wrapper if title is already italicized
in the case of book titles, which now get <em>Title</em> wrappers to render as expected, the italicized-link can 'chomp' the curly-quotes. but if the book is italicized, it's invalid style to have curly quotes anyway, so this has an easy solution of just skipping curly quotes if an em wrapper is detected.