LA: begin refactoring to pull archiving out of site compilation into a separate process; start by speeding up reads
the goal here is to start running link-archiving as a separate process, so I can rebuild the site during the day without interfering with normal work, and so start doing archiving on a daily basis to catch up. We move archiving calls out of the rewrite pass, and into a standalone function which will archive _n_ links. This standalone function can then just be called every day, at 7PM say.