Image: video posters: due to problems with the <video> tagsoup never working & switching to video include links, this code should now be all dead
we still need to *generate* poster images, however, somewhere. the TagSoup approach is very inefficient, so we won't maintain that just to do the poster generation step (and this would break on video include links anyway).  we may do this out of band during the sync-gwern.net.sh build process (since it'd be easy to just check that for every mp4/web, file, there's a corresponding -poster.jpg - fast and efficient), or possibly hook imageLinkHeightWidthSet (but lots of compile-time overhead) or something...