LM: file-transcludes: hide PDFs on mobile by default
this is a followup to ade0778ae7a3c67fc73e60849618ab1fbfb475a2; as we wound up having to pierce the encapsulation to get to the final filetype in order to write the natural language description 'click to expand PDF:' etc, we might as well statically specify that PDFs cannot be transcluded on mobile. This may permit removing the complexity of 32153a8fb5fdc79ebd997e51a4fdc9cdc0fe762e JS runtime check for '.pdf'.