YAML: move from camelCase to hyphen-separated, consistent with rest of codebase
per-page Pandoc metadata, which is in YAML, was written using camel-case like  'cssExtension' or 'safeURL', in part because I didn't know if YAML allowed hyphens in keys and didn't think to check. as more page variables accumulated, this makes them ever harder to remember in the context of all of our classes and templates, so I checked, and hyphens are safe in YAML keys, so mass rename: cssExtension->css-extension, safeURL->safe-url etc.