LM/gB: begin experimenting with removing special treatment of Twitter/WP from backend, to treat them more uniformly
the special dynamic handling of Twitter/WP will now instead be the responsibility of the frontend; it will not be implied by .link-annotated; instead, now, when a Twitter/WP link has .link-annotated or .link-annotated-partial set on it, there will in fact be a HTML annotation snippet written out & available for it just like every other .link-annotated{-partial} link. this will help with the future author-link feature (where we want to track backlinks to WP articles, because many author-links will link to their WP bio, and not to a non-WP URL).