migrate all Twitter.com -> X.com given official Musk announcement <https://x.com/elonmusk/status/1791351500217754008> & the hard redirects now imposed on twitter.com URLs
As Musk has a strange obsession with 'X' and has been ramming this through for years now despite it being a branding disaster, there is no hope he's going to give up on it and revert x.com->twitter.com /sigh. So might as well make a clean thorough break.