Test: URLs: add a hybrid to check lists of local or external URLs
motivated by Config.MetadataAuthor: instead of linking myself as 'https://gwern.net/index...', it'd make more sense to link '/index', but then it breaks the testing requirement of 'all valid external URLs'. so we add a hybrid: either a valid external URL *or* a valid URI starting in '/'. and since this is more what we mean by the URI testing, we can swap the URI tests over to that too (which exposed, as usual, a few bugs in configs etc)