Link-bibs: remove all reader-visible UI use of 'Link Bibliography' in favor of just 'Bibliography'
this renaming is motivated by watching a Gwern.net reader who, despite being a techie designer, was completely blind to the 'link bibliography' even after I named it; I think it's because 'link bibliography' is a novel phrase which breaks pattern-matching, and so is hard to 'see' at all. In followup work, we will add 3 icons to try to draw attention & define the trio of backlinks+link-bibs+similar-links.