Typography: imageCaptionLinebreak: rewrite to figureCaptionLinebreak & operate at the Block/Figure level on the Figcaption instead of Image payload
The failure-to-linebreak keeps happening in annotation figures, and after digging in, it seems that it is because the annotation HTML compiles to a full <figure><img><figcaption> triplet, but the Pandoc rewrite pass is defined on the *img* only. so if run on the Image, it works, but the Image doesn't exist when that triplet is read in. It exists only as a Figure. I didn't realize this because Images get turned into Figures later.