backlinks: include _n_ in visible snippet header so readers can more easily decide if there are enough backlinks to be interesting to look at
This is more precise & legible than the alternative of an ordinal progress indicator, because most of the length of a backlink entry is in the reverse transclusion of the context, which could be screens' worth, but the progress indicator JS couldn't possibly estimate that without very expensive loading of possibly hundreds of pages/annotations in the worst case. So skip the misleading hard-to-read ordinal and just write down the _n_ (which is easy because we already pluralize the label based on the _n_ count).