deprecate 'keywords' inside annotations as waste of space, but preserve in HTML comments in case of reversal and/or for the embedder
The '[<strong>Keywords</strong>: foo, bar]' in annotations has long troubled me. Is it worth the space and overhead to preserve them, bold them, and show them by default? When they are often so bad? when there is zero consistency and no controlled vocabulary or ontology across the thousands of sources which provide 'keywords'? When the relevant keywords will usually be hyperlinked anyway and so visibly styled differently for a reader skimming? and when the other keywords do not provide a link or anything beyond a bare naked word? do we *need* old-school 'keywords' intended for 1960s style library catalogue indexing? after 5 years of annotating, I have yet to find them actually useful, and when they have been, that usually implies a tag should be created... So I think it is time to experiment with hiding them in HTML comments by default, and see if any need surfaces. (I doubt it will. Even if it does, at this point, it may well make more sense to feed into a good LLM and create a controlled tagsonomy instead...)