LLM scripts: remove the input echoing feature, because it breaks prompt caching by its nature
I only just realized the unintended side effect of the old RNN trick. Since prompt caching saves like 50% when running a lot of calls, prompt caching will likely become much more important in the future in the SaaS APIs to avoid finetuning/dynamic-evaluation while enabling multi-million token corpuses, and the quality gain is not huge, and the fact that it breaks prompt caching might be very surprising to me or any future users, this is not a good idea, so remove it systematically.