anchor-checker.php: dbohdan update: crash if input doesn't look like it parsed as valid HTML
I discovered anchor_checker.php hadn't been finding any anchor errors because I accidentally was running it only on Markdown files, and Markdown is, apparently, technically valid HTML and so the HTML parser wouldn't crash or error out on it, but will never find any broken anchors, of course. The script should definitely warn if the input is not HTML, but it's a bit tricky to decide what is 'real' HTML. Here we first try with a heuristic that any HTML file on Gwern.net should probably contain at least one hyperlink (no counter-examples immediately came to mind) and if not, then it's invalid. Plus additional formatting.