gD: revise tag-directory headers to look inside tag-directory for an 'abstract.md'
The old approach of looking inside /note/ was not working: /note/ conceptually wasn't making sense and the indirection was ugly and confusing. it makes more sense to associate the /note/tag.md as closely as possible with the tag-directory (and resolves the collision problem too). especially as it becomes clear that the /note/ idea didn't work (and the goal of 'lightweight' pages is better handled by /blog/ right now).