sync: check for the occasional edgecase of inadequately zero-padded multiple-n files
we don't actually rely on the zero-padded convention anywhere that I know of, but it is nice to have because it makes for better sorting on the commandline and for filenames to be somewhat more predictable in bulk and to have less typographical variation. This is mostly respected on Gwern.net right now, but there are a few here or there which we don't notice, so we fix those and add a lint to warn about future occurrences.