sync: rm HTML Tidy reformatting phase due to yet another bug, which is the last straw
in this case, HTML Tidy apparently injects a lot of whitespace in HTML tables, which show up when you copy-paste them. There appears to be no option of any kind in Tidy to disable this and I see no bug reports for it. This is bad because not only does it make copy-paste bad, it destroys your ability to transclude them. /epigram#dnd-alignment-table-boomer-to-zoomer cannot be transcluded, like it is on /lorem-april-fools-2026, because it violates the blockContextMaximumLength limit... even though the table is tiny, because the *added whitespace* is enormous and trips it. (Presumably all the other tables can't be transcluded either, and we just haven't run into that edge-case before.) Because HTML Tidy keeps causing problems, every few months, like clockwork, and this one appears to have no feasible work around other than maybe checking for any table in a page & skipping, I'm biting the bullet and just disabling it entirely. We'll keep Tidy as a linter, but that's it.